+++ R/area_titles.R
@@ -1,0 +1,18 @@
+#' Dataset containing FIPS codes for counties, states and MSAs.
+#'
+#' Built-in dataset to look up areas by FIPS code.
+#' Includes recent and historical FIPS codes.
+#' To access the data directly, issue the command \code{area_titles)}.
+#' \itemize{
+#'   \item \code{area_fips}: FIPS code
+#'   \item \code{area_title}: Area name
+#' }
+#'
+#' @docType data
+#' @keywords internal
+#' 
+#'
+#' @usage data(area_titles)
+#' @note Last updated 2016-09-30
+#' @format A data frame with 4723 rows and 2 variables
+"area_titles"