+++ R/size_titles.R
@@ -1,0 +1,17 @@
+#' Dataset containing size codes for US industries by size.
+#'
+#' Built-in dataset to look up industries by size.
+#' To access the data directly, issue the command \code{size_titles)}.
+#' \itemize{
+#'   \item \code{size_code}: Size code
+#'   \item \code{size_title}: Size title
+#' }
+#'
+#' @docType data
+#' @keywords internal
+#' 
+#'
+#' @usage data(size_titles)
+#' @note Last updated 2016-09-30
+#' @format A data frame with 10 rows and 2 variables
+"size_titles"